+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased <!-- [0.1.0-alpha.3] - 2025-07-28 -->
+
+### Added
+
+- Exported `HasAlpha` and `HasGray` traits in the prelude
+
 ## [0.1.0-alpha.2] - 2025-07-28
 
 ### Added